In BOOT.LISP, added the default binary type for CMULisp, x86f.
Added a file C-TYPE-UTIL, which now holds the macros that had been previously
included at the head of C-COERCE.  Made changes in the def-c-type-coercion to
eliminate some warnings.  It seems that CMULisp is much more aggressive about
eliminating dead code and warning about it, and so some of these macros have to
attempt to be as sufficiently smart as the compiler.

Changed GLI-UTIL so that the safe versions of cons-car and cons-cdr are held in
functions.  This is needed to clear tons of "dead code eliminated" warnings from
the CMU Lisp compile.

Changed underlying-def-foreign-callable from a macro to a function to prevent
CMU Lisp from seeing dead code and whining about it.

In L-CONST and L-EXPR fixed two actual bugs that CMU Lisp pointed out.  All of
them were either erroneous declare ignore calls, or else type declarations for
variables that no longer existed in the code.