These edits enable the GLT directory to be compiled without errors in CMULisp, version 18b on Linux.
boot.lisp : Added a feature to compile-glt.  If the recompile flag is given, it
deletes the binaries before beginning the recompile.  That way, if the
recompile is interrupted in any way and fails, then the next call to
compile-glt will pick up where the previous compile left off.

destruct.lisp : The supplied-p argument was not supported for keyword
arguments.  Added it.

env.lisp : The macro env-or-default was using an incorrect test to determine if
the given argument was a valid GLT environment.  Instead of checking to see if
the environment is nil or a cons (which would catch Lucid environments), make a
positive test for the structure used in GLT.  Since CMULisp uses structures,
this was fooling the test that was in there.  Also added code to eval-when
handling so that it always transforms its situation lists into the new keyword
based ANSI CL style, i.e. :compile-toplevel, :load-toplevel, and :execute.  The
old, depricated style was not supported by CMULisp.

exports.lisp : Touched to force a full recompile.

glt-math.lisp : The function ignore doesn't exist in CMULisp.  Made a minor
rewrite to work around it.

glt-prim.lisp : Added some missing type declarations.

l-expr.lisp : More changes to eval-when to force ANSI CL style situations.

macros.lisp : Reindented parse-macro while chasing down bugs in other areas.
Fixed a reference to &env in gl:defmacro to instead be &environment.  Changed
calls to eval-when to use the ANSI CL situations.

setf.lisp : Changed to use ANSI CL situations for eval-when.

types.lisp : Changed to use ANSI CL situations for eval-when.