Boot.lisp : Fixed the feature switch for CMU from cmu17 to just plain old cmu. Added a binary type declaration for MCL.
env.lisp : Fixed up gl:declaim so that declarations of type "declaration" are
passed through to the underlying Lisp.  This makes the GL extensions to Common
Lisp known to the underlying Lisp and suppresses warnings.  Also updated an
eval-when call to the ANSI CL situations.

gli-util.lisp : Added Ben's edits that implement current-directory and
change-current-directory for GL, and fixed up some cmu17 feature switches.

glt-foreign.lisp : Added stub implementations of C foreign function calling in
CMU Lisp.  They are currently switched out with the :verify switch, so we won't
get bit by these until we can verify that they work.

glt-prim.lisp : Added debugging support that can fetch the function-name from a
compiled-function object in CMU, fiddled with a declaration to avoid a spurious
CMU warning.  It seems that the chatty CMU compiler will issue a note every
time you use an abbreviated type declaration, such as (list list-var) instead
of the more formal (type list list-var).

l-trans.lisp : Added a fat-and-slow warning for cases where a call to funcall
could not be transformed into a funcall-simple-compiled-function.

system.lisp : Added an example based on the Lecho system, which we need to
resurrect as a simple example of bootstrapping an application.

types.lisp : Fixed up the subtypep check for sequence subtypes so that two
values are always returned.