exports.lisp : Add the symbol get-setf-expansion, since the ANSI CL committee, in their infinite wisdom, moved to that from the get-setf-method name.
glt-math.lisp : Work around the lack of an ignore function.  Also work around an
Allegro CL on PC bug.  If you have a macro named "defblahblah" and the first
argument it takes is a string, and you call it at top level of a file, then ACL
will croack attempting to record-source-location on the property list of the
string.  I believe that CMU has similar trouble.

glt-out.lisp : Rename the C tranlation simple-error to gli-simple-error.
Simple-error is now a Lisp package symbol, and we need to sidestep it.

glt-prim.lisp : Same problem with strings as arguments to def-like macros.
Also, changed calls to get-setf-method to get-setf-expansion.

setf.lisp : Many changes from get-setf-method to get-setf-expansion.