boot.lisp : Sp[lit the loop that actually compiles modules out of the function that compiles and loads the boot module itself.  This way, if there is a change in the way modules are loaded, a newly compiled version of that function will be run immediately after loading.  Also, set *read-default-float-format* to double-float to avoid those pesky single floats.
c-coerce.lisp : Changed some of the C type coercions to use new macros that hide
the messy details of some of these conversions.  Before this edit, all type
conversions were explicitly performed in open code so as to hide nothing about
the performance implications of type changes.  However, most of these
conversions are 2 to 3 instruction operations, and so wrapping them up in a
macro is not too unreasonable, and it hides a lot of cruft.  Besides the new
BOXFIX, UNBOXFIX, BOXCHAR, UNBOXCHAR, ObjStrHDR, and ObjSvHDR macros that now
encapsulate previously existing transforms, there are new transforms StrHDR and
SvHDR that take the unboxed arrays and return pointers to their structure
types.  Previously these were transformed to type Obj and then had to be cast
back into the appropriate structure types.

c-names.lisp : Added new name reservations for the new macros StrHDR, ObjStrHDR,
SvHDR, and ObjSvHDR.  Also, the C identifier for "main" hadn't been reserved,
thought I can't believe it through the testing its experienced so far.

exports.lisp : Fiddled with the copyright date and name.  Did it as a gratuitous
touch of the file, because all of this will change once we've gone open source.

gli-util.lisp : Added an implementation for Allegro to gc-a-little.  Also, added
a function called gc-a-lot, which does a full garbage collect.

glt-out.lisp, glt-prim.lisp, l-trans.lisp : Took advantage of the new type
coercions to pointer to Str and pointer to Sv when fetching lengths and
fill-pointers.

system.lisp : Added new functions that return the makefile locations for a given
system.  Also added a function that takes two pathnames and prints to a given
string the pathname that can be used to cd from the first to the second.

trandata.lisp : Added an :if-exists :supersede flag to the code for opening up
the translation data file.  This prevents a crash in ANSI compliant Lisps when
performing a second translation of a file.

trans.lisp : Added :supersede options to the remaining file open for output
function calls.  Also, call the function that generates makefiles.