boot.lisp : Added counts to the display when compiling GLT.
c-coerce.lisp, decls.lisp, glt-prim.lisp, types.lisp : Changed the runtime
implementation type for string-streams to be the structure type gl-string-stream
instead of cl:string-stream.  Lucid had not implemented string-streams, so we
were able to use that symbol as a GL type without collision.  However, when
compiling in CMU, it croaked when it hit the conflict that what it meant by
string-stream was different from what we meant.  So, GL is now free of
references to cl:string-stream, and gl:string-stream expands into
gli::gl-string-stream.  Got that?