Defun.lisp : Added some printing code for debugging.  It's currently commented out.
Special.lisp : Added new clauses for expression-result-type.  I had been relying
on constantp returning T for structure instances, but the definition of constantp
explicitly gives leeway on this operation.  Changed it to use atom, and reorganized
the code a little.

Types.lisp : Added a default initialization value for variables of type character.
In this case, that variable gets a null character (value of 0).