Add a mechanism that enables the GL user to customize what appears in the copyright placed into the translated C sources, i.e. a function is invoked to generate that copyright, the function gets a mess of into it can use at it's option when decorating the code.  This is part of the copyright rework project.
Add another makefile target, in this case "conf", for the files that
input to the autoconf makefile mechinism.  If this were to have legs
then it might displace the current one per platform style.  This
works for me, but I don't think it's quite featured up enough to
work for any interesting examples.  There are design tangles here.
The user probably wants to define a system as a tree of systems
and then get rewarded with something he can kick a little and then
wrap up in a tar file and ship to a user who then types
"./configure & make & make install".  That "kick a little" step
is currently far to complex.  This is part of the trophy wife
(she's thin, she's round all over, she's smart, she's fast) example
application project.