tlt/lisp/tli-util.lisp : Added some extra force-output calls to the translation-error and translation-warning systems.  Without them, in CMU on Linux the lines naming the current context (i.e. function name or top-level form) are printed after the first warning message for that function instead of before.  Also added CMU implementations for gc-a-little and gc-a-lot.
tlt/lisp/l-expr.lisp : Added a hook to check that there is a currently defined
area when calling a function declared as a conser, i.e. one that requires the
calling context to establish either a permanent or a temporary area before it's
called.

tlt/lisp/l-trans.lisp : Fixed a typo in a comment.

tlt/lisp/makefiles.lisp : Fixed up an indenting error in a comment.

tlt/lisp/regions.lisp : Added a function `check-area-for-conser-call' which is
responsible for checking for proper lexically apparent consing declarations when
a call is make to a "conser" function.  If there is no area declared, and there
is a current scope-name, then an error will be signalled.  Scope names are
visible when we are within the lexical scope of a function, instead of a
top-level load-time eval form.  Top level forms are in the permanent area by
default, and so do not need to be declared.