This commit implements deep binding for special variables when we're compiling with the make line "make THREAD=1".
In the directories lecho/c and tl/c, many translated C files had differences due
to the changes in how global variable references and sets are translated.

tl/c/tl.c : Made many minor changes to the creation and current_thread_state
functions, including the addition of a current_thread_state_if_any() function
to be used when looking up global variable values.  The major change was the way
that the thread_states_mutex is locked, to prevent the creation of a new array
of thread states from freeing an old array which it is still being referenced in
another thread.  Added functions for binding and unbinding special variables,
for getting the address of the current binding for a special, for fetching and
setting the value of a special, and to use a new enumeration type for stack
frame type markers, rather than the magic constants that were previously used.

tl/c/tl.h : Added the Stack_frames enumeration, removed all references to the
now unneede Binding structure.  Added macros for getting and setting global
variable values.  Note that the PTHREAD compiler switch will now toggle special
variable handling back and forth between shallow binding and deep binding.

tlt/lisp/boot.lisp : Added a new file, l-stack.lisp.

tlt/lisp/c-coerce.lisp : Added a new def-c-type for the thread-state structure,
so that compilations can cache the current thread state and limit the number of
times it must be looked up.

tlt/lisp/c-names.lisp : Added the names of global variable lookup functions to
the list of reserved identifiers.

tlt/lisp/c-types.lisp : Added documentation for the thread-state type.

tlt/lisp/l-trans.lisp : Made changes to deal with the different translation
needs for special variables within threads.  In particuar, the functions
bind-global-for-let, unbind-global-for-let, and their associated helper
functions that take lists of bindings have been moved to the new l-stack.lisp
file.  There they were rewritten to call the functions that implement binding
and unbinding, rather than open code the implementation.  Perhaps some day these
binding functions should be made into macros to regain the previous
performance.  The translations for let and let* were changed to allow the
current thread state to be cached so that it wouldn't have to be looked up more
than once per LET.  The translations for setq and implicit-symbol-value were
changed to use the new GET_GLOBAL and SET_GLOBAL macros when referencing special
variables.