In defstruct.lisp, finished the implementation of reclaimers on structures.   Now there are working reclaimer options, working slot reclaimer options, and working slot copier options.  Also hushed up a leak associated with the funcallable slot accessor functions of type declared structure elements.  It still leaks, but the accessor is declared with a consing type of "either", which makes the accessor no longer issue compiler warnings.  Note that this leak can only occur when you actually funcall (or apply) the accessor.  If you call it normally the compiler macro will inline the accessor, in which case the consing characteristics are the responsibility of the calling location.
Iin tli-util.lisp, removed extra arguments to gc within the gc-a-lot function
for CMU Lisp.  I'm using the released version of CMU Lisp, which doesn't have the
generational garbage collector, and so gc doesn't have the arguments to force a full
gc.  Maybe when those guys actually release something we can put it back in.  -jallard