Edits to tlt/lisp files as part of a port to CLISP.
boot.lisp : Turn off the "*package-tasks-treat-specailly*" feature of CLISP.

l-const.lisp : Changed the handling of fixnums and integers in the code that
chooses types while compiling references to constants.  Since CLISP and ThinLisp
have different fixnum sizes, check all integers for fixnum emitting, not just
the underlying Lisp's idea of what a fixnum is.

l-top.lisp : Bind print-pretty to NIL to eliminate unwanted newlines in emitted
C files.

l-trans.lisp : Bind print-pretty to NIL to eliminate unwanted newlines in
emitted C files.

system.lisp : Change the output style for strings while compiling.  When using
CLISP in a shell buffer below emacs, if you use format each character is output
separately and you can see each character going out.  By using write-string
after formatting each line, it snaps onto the screen in a pleasing way.  (Ok,
this was a dumb edit, but I like it.)

trandata.lisp : Bind print-pretty to NIL to prevent unwanted newlines in tlt
files.

trans.lisp : More binding of print-pretty and buffering of output.  Also changed
to not rely on rename-file to overwrite an existing file in the target of the
rename.  Now we explicitly delete any file before overwriting it.

types.lisp : Changed tl-type-p and tl-type-of handling for fixnums to use a
bounds check on the ThinLisp idea of fixnum sizes, not to use the underlying
Lisp.