Improvements to Common Lisp coverage.
In stubs.lisp, changed the stub for coerce to at least handle the cases where
the value is being coerced to a double-float.  Removed the stubs for functiop
and assert, since translatable implementations of those operations have been
added.

In tl-types.lisp, added type defintions for single-float, short-float, and
long-float to make them all equivalent to double-float.  Added a definition for
functionp that checks if the given value is a compiled-function.  This is
adequate since we don't have interpreted functions in ThinLisp.

In tl-util.lisp, added a definition for assert.  Since we don't have continuable
errors, this implementation of assert just calls error.