Changed the Lisp macro expand time evaluations of the accessors for compiled-function objects to consistently use the derror form.  This is necessary to eliminate dead code warnings in CMU Lisp.  Here's why.
In CMU Lisp if a function or macro expansion can be determined to return a type or
a constant then the compiler will use that value to further constant fold any uses
of the result of that function.  When these constant folded values are part of the
predicate argument to an IF or COND form, it will eliminate the dead code that is
part of the consequent clause that won't be called.  This is all well and good, but
then the CMU compiler proceeds to whine that it is eliminating dead code.

The reason this comes up is that some of the operations defined by def-c-translation
do not have a Lisp development time equivalent.  The compled-function-sets-values-count
accessor of ThinLisp compiled functions is a case in point.  Since there is no Lisp
time evaluation that would make sense, if this code is executed during Lisp
development, the expansion will call error.  CMU Lisp was seeing this call to
error and eliminating all the code following that point in the function flow graphs
where it was being called.  This was then causing these spurious dead code warnings.
So, this edit calls derror instead of error.  Derror is written in a way to
confuse CMU's code analyzer so it will stop making assumptions and bothering me with
compiler warnings.

This edit uses derror more to eliminate dead code warnings that I was getting in
the compilation of tl/lisp/apply.lisp.